+++ tests/unit/classes/classes/BazClass.php
@@ -2,6 +2,9 @@
 
 namespace mako\tests\unit\classes\classes;
 
+use AllowDynamicProperties;
+
+#[AllowDynamicProperties]
 class BazClass
 {
 	use FooTrait;